requirements.txt
opencv-python
pyautogui
pygetwindow
numpy
mss
pillow
loguru
python-dotenv
